requirements.txt
discord
python-dotenv~=0.19.2
requests
nextcord~=2.0.0a6
psycopg2~=2.9.3
dhooks~=1.1.4
yfinance~=0.1.69
googlefinance~=0.7
iexcloud~=0.0.1